feat(EG-420): separate auth utility functions from rest-api-utils.ts to auth-utils.ts
