+++ packages/back-end/src/app/utils/rest-api-utils.ts
@@ -1,9 +1,3 @@
-import {
-  LaboratoryAccess,
-  LaboratoryAccessDetails,
-  OrganizationAccess,
-  OrganizationAccessDetails,
-} from '@easy-genomics/shared-lib/src/app/types/easy-genomics/user';
 import { APIGatewayProxyEvent } from 'aws-lambda/trigger/api-gateway-proxy';
 
 export const enum REST_API_METHOD {
@@ -80,47 +74,4 @@ export function getApiParameters(event: APIGatewayProxyEvent): URLSearchParams {
   }
 
   return parameters;
-}
-
-/**
- * Helper function to check if an authenticated User's JWT OrganizationAccess
- * metadata allows access to the requested Organization / Laboratory.
- * @param event
- * @param organizationId
- * @param laboratoryId
- */
-export function validateOrganizationAccess(
-  event: APIGatewayProxyEvent,
-  organizationId: string,
-  laboratoryId?: string,
-): Boolean {
-  try {
-    const orgAccessMetadata: string | undefined = event.requestContext.authorizer?.claims.OrganizationAccess;
-    if (!orgAccessMetadata) {
-      return false;
-    }
-
-    const organizationAccess: OrganizationAccess = JSON.parse(orgAccessMetadata);
-    const organizationAccessDetails: OrganizationAccessDetails | undefined = organizationAccess[organizationId];
-    if (!organizationAccessDetails || (organizationAccessDetails && organizationAccessDetails.Status !== 'Active')) {
-      return false;
-    }
-
-    if (laboratoryId) {
-      const laboratoryAccess: LaboratoryAccess | undefined = organizationAccessDetails.LaboratoryAccess;
-      if (!laboratoryAccess) {
-        return false;
-      }
-
-      const laboratoryAccessDetails: LaboratoryAccessDetails | undefined = laboratoryAccess[laboratoryId];
-      if (!laboratoryAccessDetails || (laboratoryAccessDetails && laboratoryAccessDetails.Status !== 'Active')) {
-        return false;
-      }
-    }
-
-    return true;
-  } catch (error) {
-    console.error(error);
-    return false;
-  }
 }
